pdms_app/serializers.py
from rest_framework import serializers
from .models import Task, UploadedCSV

class TaskSerializer(serializers.ModelSerializer):
    class Meta:
        model = Task
        fields = ['id', 'title', 'description', 'assigned_to']

class UploadedCSVSerializer(serializers.ModelSerializer):
    class Meta:
        model = UploadedCSV
        fields = ['id', 'file', 'uploaded_at']
