requirements.txt
asgiref==3.11.1
charset-normalizer==3.4.4
Django==6.0.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
lxml==6.0.2
pillow==12.1.1
PyJWT==2.11.0
python-docx==1.2.0
reportlab==4.4.10
sqlparse==0.5.5
typing_extensions==4.15.0
tzdata==2025.3
